museum mode with chat added
