+++ client/src/components/layout.tsx
@@ -1,4 +1,4 @@
-import { useEffect } from "react";
+import { useCallback, useEffect } from "react";
 import ClickGate from "./gates/globalClickedGate";
 import Footer from "./footer";
 import useMediaQuery from "../stores/useMediaQuery";
@@ -8,15 +8,18 @@ import Instrumentation from "./Instrumentation";
 import GlobalRoomsSummaryProvider from "./gates/globalRoomsSummaryProvider";
 import GlobalPresenceGate from "./gates/globalPresenceGate";
 import GlobalUserAdminProvider from "./gates/globalUserAdminStore";
+import { useMuseumMode } from '../stores/useMuseumMode';
+import MuseumFooter from './MuseumFooter';
 
 
 //TODO: Currently we have a listener for each ringNode. Can simplify, requires DB refactor.
 
 const Layout: React.FunctionComponent<{
   hideChat?: boolean;
-  hideFooter?: boolean;
-}> = ({ children, hideFooter }) => {
+}> = ({ children }) => {
   const isMobile = useMediaQuery();
+  const isMusuemMode = useMuseumMode(useCallback(s => s.isMuseumMode, []));
+
   useEffect(() => {
     const setViewportHeight = () => {
       const vh = window.innerHeight * 0.01;
@@ -31,23 +34,29 @@ const Layout: React.FunctionComponent<{
 
   return (
     <ClickGate>
-    <div
-      className={classnames("fullScreen lightFill relative noOverflow", { "stack:noGap": isMobile })}
-      key={`layout-${isMobile}`}
-    >
-      <Head>
-        <meta name="viewport" content="width=device-width, initial-scale=1.0, maximum-scale=1.0, user-scalable=no" />
-      </Head>
-        <GlobalRoomsSummaryProvider>
-          <GlobalUserAdminProvider>
-            <GlobalPresenceGate>
-              {children}
-              {!hideFooter && <Footer />}
-            </GlobalPresenceGate>
-          </GlobalUserAdminProvider>
-        </GlobalRoomsSummaryProvider>
-      <Instrumentation />
-    </div>
+      <div
+        className={classnames("fullScreen lightFill relative noOverflow")}
+        key={`layout-${isMobile}-${isMusuemMode}`}
+      >
+        <div className={classnames({
+          "fullScreen relative": !isMusuemMode,
+          "stack:noGap": isMobile || isMusuemMode,
+          "fourByThree": isMusuemMode
+        })}>
+          <Head>
+            <meta name="viewport" content="width=device-width, initial-scale=1.0, maximum-scale=1.0, user-scalable=no" />
+          </Head>
+          <GlobalRoomsSummaryProvider>
+            <GlobalUserAdminProvider>
+              <GlobalPresenceGate>
+                {isMusuemMode ? <div className="sixteenByNine">{children}</div> : children}
+                {isMusuemMode ? <MuseumFooter /> : <Footer />}
+              </GlobalPresenceGate>
+            </GlobalUserAdminProvider>
+          </GlobalRoomsSummaryProvider>
+          <Instrumentation />
+        </div>
+      </div>
     </ClickGate>
   );
 };
